Add Routes and Links for nav
