+++ src/App.js
@@ -1,42 +1,17 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
+import { Route } from "react-router-dom";
+
 import styled from "styled-components";
 import Nav from "./components/Nav";
+import Home from "./components/Home";
+import Footer from "./components/Footer";
+
 import Intro from "./components/Intro";
 import Summary from "./components/Summary";
 import Skills from "./components/Skills";
 import Projects from "./components/Projects";
 import Contact from "./components/Contact";
 import ContactCompleted from "./components/Contact/ContactCompleted";
-import Footer from "./components/Footer";
-
-import { Transition } from "react-transition-group";
-import Lottie from "react-lottie";
-import animationData from "./theme/bow.json";
-
-// transition
-const duration = 450;
-
-const defaultStyle = {
-  transition: `opacity ${duration}ms ease-in-out`,
-  opacity: 0,
-};
-
-const transitionStyles = {
-  entering: { opacity: 1 },
-  entered: { opacity: 1 },
-  exiting: { opacity: 0 },
-  exited: { opacity: 0 },
-};
-
-// animation
-const defaultOptions = {
-  loop: false,
-  autoplay: true,
-  animationData: animationData,
-  rendererSettings: {
-    preserveAspectRatio: "xMidYMid slice",
-  },
-};
 
 //////////// STYLED COMPONENTS /////////////
 const AppDiv = styled.div`
@@ -61,6 +36,12 @@ const DIVIDE = styled.div`
   margin: 6rem auto;
 `;
 
+const SPACE = styled.div`
+  max-width: 82.5rem;
+  width: 95%;
+  margin: 6rem auto;
+`;
+
 const FooterContainer = styled.footer`
   width: 95%;
   margin: 0 auto;
@@ -70,85 +51,41 @@ const FooterContainer = styled.footer`
 //////////// REACT COMPONENT /////////////
 export default function App() {
   const [submitted, setSubmitted] = useState(false);
-  const [pageIn, setPageIn] = useState(false);
-  const [svgIn, setSvgIn] = useState(true);
-
-  useEffect(() => {
-    // remove scrollbar during anmiation
-    const body = document.querySelector("body");
-    body.style.overflow = "hidden";
-
-    const timer = setTimeout(() => {
-      // trigger fade-out
-      setSvgIn(false);
-      body.style.overflow = "auto";
-      // then unmount
-      unMountSVG();
-    }, 4000);
-    return () => clearTimeout(timer);
-  }, []);
-
-  const unMountSVG = () => {
-    setTimeout(() => {
-      console.log("Mounting home page");
-      setPageIn(true);
-    }, 310);
-  };
 
   return (
     <AppDiv>
       <Header>
         <Nav />
       </Header>
-      {!pageIn ? (
-        <Transition in={svgIn} timeout={0}>
-          {(state) => (
-            <>
-              <Lottie
-                options={defaultOptions}
-                height={"70vh"}
-                width={180}
-                style={{
-                  ...defaultStyle,
-                  ...transitionStyles[state],
-                }}
-              />
-              <FooterContainer>
-                <Footer />
-              </FooterContainer>
-            </>
-          )}
-        </Transition>
-      ) : null}
 
-      <Transition in={pageIn} timeout={0}>
-        {(state) => (
-          <main
-            style={{
-              ...defaultStyle,
-              ...transitionStyles[state],
-            }}
-          >
-            <Intro />
-            <Skills />
-            <Summary />
-
-            <DIVIDE />
-
-            <Projects />
-
-            <DIVIDE />
-
-            {submitted ? (
-              <ContactCompleted />
-            ) : (
-              <Contact setSubmitted={setSubmitted} />
-            )}
-
-            <DIVIDE />
-          </main>
+      <Route exact path="/">
+        <Home
+          submitted={submitted}
+          setSubmitted={setSubmitted}
+          DIVIDE={DIVIDE}
+          FooterContainer={FooterContainer}
+        />
+      </Route>
+      <Route path="/about">
+        <Intro />
+        <Skills />
+        <Summary />
+        <DIVIDE />
+      </Route>
+      <Route path="/projects">
+        <SPACE />
+        <Projects />
+        <DIVIDE />
+      </Route>
+      <Route path="/contact">
+        <SPACE />
+        {submitted ? (
+          <ContactCompleted />
+        ) : (
+          <Contact setSubmitted={setSubmitted} />
         )}
-      </Transition>
+        <DIVIDE />
+      </Route>
 
       <FooterContainer>
         <Footer />
