+++ src/data/skills.json
@@ -1,1 +1,1 @@
-["JavaScript", "ReactJS", "NodeJS", "Python3", "Java Spring", "CSS"]
+["JavaScript", "ReactJS", "NodeJS", "Python3", "Java", "CSS"]
